+++ crypto-common/src/commonMain/kotlin/models/CryptoPairRate.kt
@@ -1,6 +1,8 @@
 package models
 
+import java.math.BigDecimal
+
 data class CryptoPairRate(
     var pair: CryptoPair = CryptoPair(),
-    var rate: Double = 0.0
+    var rate: BigDecimal = BigDecimal.valueOf(0.0)
 )
